final changes, cleaned up some stuff and made sure it works with just one card sub
